refactor: 빈 설정오류 수정. 불필요 코드 제거
