+++ src/main/java/kr/hhplus/be/server/ServerApplication.java
@@ -3,10 +3,11 @@ package kr.hhplus.be.server;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
-//@MapperScan(value = {"kr.hhplus.be.server.user.repository", "kr.hhplus.be.server.point.repository", "kr.hhplus.be.server.goods.repository", "kr.hhplus.be.server.order.repository"})
 @SpringBootApplication
 @MapperScan("kr.hhplus.be.server.**.infrastructure.persistence.mapper")
+@ComponentScan(basePackages = {"kr.hhplus.be.server.*"})
 public class ServerApplication {
 
 	public static void main(String[] args) {
