+++ src/main/java/kr/hhplus/be/server/order/domain/service/OrderIdGenerator.java
@@ -1,0 +1,14 @@
+package kr.hhplus.be.server.order.domain.service;
+
+import org.springframework.stereotype.Component;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+@Component
+public class OrderIdGenerator {
+    public String generate() {
+        String now = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMddHHmmssSSS"));
+        return "E" + now;
+    }
+}
