+++ src/main/resources/application.yml
@@ -2,7 +2,7 @@ spring:
   application:
     name: hhplus
   profiles:
-    active: local
+    active: test
   datasource:
 #    name: HangHaePlusDataSource
     url: jdbc:mysql://localhost:3306/ecommerce?characterEncoding=UTF-8&serverTimezone=UTC
